Issue #1315 - plugin utilities.
