+++ libraries/plugins/account_history/account_history_plugin.cpp
@@ -6,6 +6,7 @@
 
 #include <steemit/chain/operation_notification.hpp>
 #include <steemit/chain/history_object.hpp>
+#include <steemit/utilities/json_operations.hpp>
 
 #include <fc/io/json.hpp>
 #include <fc/smart_ref_impl.hpp>
@@ -14,20 +15,6 @@
 
 
 #define STEEM_NAMESPACE_PREFIX "steemit::protocol::"
-
-template<typename T>
-T dejsonify(const std::string& s) {
-   return fc::json::from_string(s).as<T>();
-}
-
-// TODO: Move this somewhere else. Also exists in app/plugin.hpp, which will be removed.
-#ifndef STEEM_LOAD_VALUE_SET
-#define STEEM_LOAD_VALUE_SET(options, name, container, type) \
-if( options.count(name) ) { \
-   const std::vector<std::string>& ops = options[name].as<std::vector<std::string>>(); \
-   std::transform(ops.begin(), ops.end(), std::inserter(container, container.end()), &dejsonify<type>); \
-}
-#endif
 
 namespace steemit { namespace plugins { namespace account_history {
