+++ libraries/plugins/witness/witness_plugin.cpp
@@ -33,6 +33,7 @@
 #include <steemit/chain/index.hpp>
 
 #include <steemit/utilities/key_conversion.hpp>
+#include <steemit/utilities/json_operations.hpp>
 
 #include <fc/io/json.hpp>
 #include <fc/smart_ref_impl.hpp>
@@ -68,20 +69,6 @@ void new_chain_banner( const steemit::chain::database& db )
       "\n";
    return;
 }
-
-template<typename T>
-T dejsonify(const string& s) {
-   return fc::json::from_string(s).as<T>();
-}
-
-// TODO: Move this somewhere else. Also exists in app/plugin.hpp, which will be removed.
-#ifndef STEEM_LOAD_VALUE_SET
-#define STEEM_LOAD_VALUE_SET(options, name, container, type) \
-if( options.count(name) ) { \
-   const std::vector<std::string>& ops = options[name].as<std::vector<std::string>>(); \
-   std::transform(ops.begin(), ops.end(), std::inserter(container, container.end()), &dejsonify<type>); \
-}
-#endif
 
 namespace detail {
